+++ resources/client_osx.rb
@@ -1,0 +1,7 @@
+unified_mode true
+
+provides :git_client, platform: 'mac_os_x'
+
+action :install do
+  package 'git'
+end