+++ spec/unit/recipes/source_spec.rb
@@ -5,8 +5,7 @@ require 'spec_helper'
 describe_recipe 'git::source' do
   context 'when using include_recipe or adding git::default to the run_list' do
     it 'installs git_client[default]' do
-      expect(chef_run).to install_git_client('default')
-        .with(provider: Chef::Provider::GitClient::Source)
+      expect(chef_run).to install_from_source_git_client('default')
     end
   end
 end