+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable `unified_mode` for Chef 17 compatibility
 - Add some missing test coverage
 - Convert `git_service` LWRP to custom resource
+- Convert `git_client` LWRP to custom resource
+  - **BREAKING**: source provider changed to `:install_from_source` action
 
 ## 10.1.0 (2020-10-12)